split format and parse
