+++ src/format.spec.ts
@@ -459,66 +459,4 @@ describe('format', () => {
       ); // This is to check for the code with date pre-2018
     }, `TypeError: Lookup functions do not currently support the provided date '2017-12-31T23:59:00.000Z'. Support is available for dates starting from 2018 onwards.`);
   });
-
-  it('parse will do something', () => {
-    assert.deepEqual(formatLibrary(at).parse('0', 'USD'), {
-      amount: 0n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('-0', 'USD'), {
-      amount: 0n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('0.01', 'USD'), {
-      amount: 1n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('-0.01', 'USD'), {
-      amount: -1n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('-1.2', 'USD'), {
-      amount: -120n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('-$1.23', 'USD'), {
-      amount: -123n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('$1.23', 'USD'), {
-      amount: 123n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('-USD$1.23', 'USD'), {
-      amount: -123n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('$-1.23', 'USD'), {
-      amount: -123n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('USD$-1.23', 'USD'), {
-      amount: -123n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('US$123456', 'USD'), {
-      amount: 12_345_600n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('US$123456.', 'USD'), {
-      amount: 12_345_600n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('US$123456.7', 'USD'), {
-      amount: 12_345_670n,
-      currency: 'USD',
-    });
-    assert.deepEqual(formatLibrary(at).parse('US$123,456.78', 'USD'), {
-      amount: 12_345_678n,
-      currency: 'USD',
-    });
-    assert.throws(() => formatLibrary(at).parse('123456.789', 'USD'), {
-      message: 'decimalPlaces > minorUnitDigits',
-    });
-  });
 });
