+++ src/index.ts
@@ -9,13 +9,16 @@
 import currencies, { type CurrencyLibrary } from './currency.ts';
 import countries, { type CountryLibrary } from './country.ts';
 import format, { type FormatLibrary } from './format.ts';
+import parse, { type ParseLibrary } from './parse.ts';
 
 export type * from './currency.ts';
 export type * from './country.ts';
 export type * from './format.ts';
+export type * from './money.ts';
+export type * from './parse.ts';
 
 export default function (
   at: string,
-): CurrencyLibrary & CountryLibrary & FormatLibrary {
-  return { ...countries(at), ...currencies(at), ...format(at) };
+): CurrencyLibrary & CountryLibrary & FormatLibrary & ParseLibrary {
+  return { ...countries(at), ...currencies(at), ...format(at), ...parse(at) };
 }
